Status of limit switches updated independent of movement direction. Issue #35
The raw limit switches found in the MSTA field are converted into
LLS and HLS:
- When the DIR field is 1, they are swapped
- When the limit switch is not in the direction of the
  commanded direction, it is not shown.
  This is useful when a controller has only one bit which reports
  both limit switches.
  It had been useful when these controllers where more common.
  Today there are different reasons to remove this filtering
  from the record:
  - the motor can be moved outside the record into the other direction
  - The motor is slowly creeping towards the limit switch and hits it
  - The cable is broken or disconnected
  - The HOMF button is pressed and the motor starts his own homing sequence,
    which hits the low limit switch.

In any case, this kind of filtering can be done in the driver, if needed.
When hitting a limit switch while doing retries or backlash, the filter
is still used to stop (or not) the motor.
(I think that it would be safer to stop the motor whenever any limit switch
 is hit, but that is a seperate discussion, don't change this now)

Conflicts:
        axisApp/AxisSrc/axisRecord.cc