Rename motorRecord into axisRecord
Do a big rename of the record, including all databases.
The rename is done mostly automatically.
asynMotorAxis.cpp is now asynAxisAxis.cpp, friend to asynAxisController.cpp

MotorSrc/ is removed and the remaning code is in AxisSrc/
The *Util*.cc files are updated and commented out in the Makefile.

Note:
If you want to see the history of a file including the time it spend under
MotorSrc, used "git log --follow motorApp/AxisSrc/axisRecord.dbd"

All drivers providing dev/drv support are commented out from the Makefile.
They will be removed one-by-one in the next commits.
If needed, they can be restored, one-ny-one, on a branch.
Use  "git revert" and use the code to write a model 3 driver.
Once tested, a pull request can be done.

Remove motor_files.html, there are no releases yet.

Note: motorRecord.html needs to be renamed and updated