Add simulator
Add a simulator for a motion controller talking TCP/IP.
This is a replacement for the old simulator, mainly for
training and testing.
The simulator has 2 frontendd: the commands we use for thr
EtherCAT based systems, the IcePAP from ESRF and the 3rd one
is disabled right now.
Today the commands don't overlapp, and the simulator accepts both
labuages.

It may be wise to select only one protocol in the future.