Add Makefile for ESS EPICS Environment EEE
The EPICS version at ESS allows different versions of EPICS base,
modules (in our case asyn) to be compiled and installed side-by-side.

The make system is based on a version from PSI and has been presented here:
https://indico.esss.lu.se/event/507/session/2/contribution/54

In order to support the "classic" EPICS build systems and EEE, 2 different
Makefiles are used: Makefile.epics and Makefile.EEE.

The new Makefile is now a wrapper which decides on environment variables
EPICS_ENV_PATH, EPICS_MODULES_PATH and EPICS_BASES_PATH what to do.

This should work for all installations, if not, we will find a solution.