.gitignore: Don't ignore axisApp/Db/
The source of db (or template) files is under ./axisApp/Db/, and
when installing, they are copied into ./db/

Therefore the /db/ directory is ignored.
On a case insensitive file system even Db/ is ignored, and new files
under ./axisApp/Db/ are not detected by Git under Mac or Windows.

Use the correct path specifier "/db/".