Add field MsgTxt in driver
Add a text field to show what the motor is doing.
The base idea is taken from motor8x.opi, which has a number of texts
displayed like "Moving" or "Soft limit".
On the long run, it is better to prepare these texts already in the IOC,
and use which display whatever to display it, including camonitor.

The current implementation does not (yet) show soft limit violations,
like the widget does.

The main advantage is that the driver can put in human readable texts,
which are fetched from the motion controller.

This is demonstrated in the Ethercat controller and simulator.

Add opi files for Ethercat, especially to show the MsgTxt

Step the version to 10.0.1 (or 10.000001)