TC 110: Check that motor is "unhomed"
Improve the homing test:
Wait for start, and check that the motor has started:
when sitting on a limit switch, the motor will never start.
This is not tested now, but more tests may be added later.
Note: The polling has a risk to fail, so comment it out for the moment.
As we test homing against LLS -> homing against HLS, this risk is low.

Correct syntax errors of code that is not (yet) uesed.