EthercatMC: Send acceleration and deceleration to the MCU
This reverts commit 1c2c4c05,
 "EthercatMC: Don't send acceleration from record to MCU"
and sends both acceleration and deceleration to the EtherCAT based
controllers.

The main reason is that the record interface expects the axis to perform
a trapezoidal movement, and we want to follow this convention, as users
may expect it to work.
This opens the door for 2 different ways to configure acceleration:
a) in the record, as users may expect things to work.
b) setting the fields in the record to 0, and configure the controller
   in st.cmd

For the moment we go for a).