Record: print name and reason when axis is stopped
Unify the printf() when the record stops the axis.
We have one of 3 outputs:
When the "New Target Monitor" decides to stop:
"STOP %s diff=%g ntm_deadband=%g"

When we are close to the soft limit and decide to stop:
"STOP %s lvio"

When either the limit switch in travelling direction is activated
or the PROBLEM bit is set:
"STOP %s ls_active=%d PROBLEM=%d"

In all cases %s shows the PV name.