Make "stop if driver returns RA_PROBLEM true" optional.
Commit 303a9208e368ef07839afff67 introduced the feature to stop
a motor when the driver sets the PROBLEM bit.

This has been problematic for different reasons:
- The commit message does not mention which controller/driver
  actually need this feature.
- The driver itself could have been fixed to do the stop
- Some controllers (like XPS) report a PROBLEM when the axis
  is not homed.
- Other controllers report errors and stop the motor on their own.
  But as soon as the axis is stopped, the error disappears and
  there is no chance for the user to see it.
  (This happens e.g. when an axis is below the low soft limit
   and a TWF is done. The record allows this, but the controller
   does not, as the new set point is out of range)

After some forth and back changes I come to the conclusion that this
feature is worth to keep, but the driver must enable it.
In other words, it does what versions of motorRecord < 6.9 did.