EthercatMC: Make it possible to send strings to the simulator
Add a new template,  axisApp/Db/EthercatMCdebug.template
Using e.g.
"caput IOC:m1-DbgStrToMCU Main.M1.bEnable=0"
it is possible to switch off the amplifier behind the back of the
record and model 3 driver.

In order to avoid to writ to any axis, there is a list of allowed commands:

1) Commands to the simulator:
  caput IOC:m1-DbgStrToMCU Sim.M1.dbgOpenLogFile=M1.log
  caput IOC:m1-DbgStrToMCU Sim.M1.dbgCloseLogFile

2) Commands to Main:
  caput IOC:m1-DbgStrToMCU Main.M1.bEnable=0

3) ADR commands using integer:
  caput IOC:m1-DbgStrToMCU "setADRinteger 501 0x4000 0x15 1"

4) ADR commands using floating point (double)
  caput IOC:m1-DbgStrToMCU "setADRdouble  501 0x4000 0x6 0.1"

Important note:
In one of the next commits a range check must be implemented,
to make sure that e.g. "M1" is only send to axis 1.

How to use this and load the data base:
See test/startup/st.SimAxis.cmd