auto-generate documentation/axisRecord.html
The current documentaton of record fields was manually maintained is today not
100% up to date.
Take a first step towards an auto-generated html file:
Save the .dbd file in a .dbd.pod file and convert it into the dbd and html
using gen_axisRecord_html.sh in the spirit of EPICS Base R3.15.4 and higher.

One problem needs to be overcome: We want to be able to use the axisReord with
older EPICS base vesions.
Since we don't expect changes to the .dbd.pod file that often, the resulting
.pod file is saved in Git as well as the html file.
In other words, if the .dbd.pod file is changed, it must be done against
EPICS base R3.15.4 or higher.