Fast STUP & SYNC did not work
When the VAL field should by synchronized with the RBV field
(similar for DVAL and RVAL) the SYNC field can be used.

When a write to STUP is followed by a write to SYNC, the the
status update may still be in progross when the SYNC is handled.
The update does not work as expected.

Solution: Ignore SYNC as long as STUP is not OFF.
Once the status update is completed, execute the SYNC.