EthercatMCAxis: bBusy handling depending on the MCU
The Twincat verion of the MCU reports bBusy so that it can
"nearly" used for >motorStatusDone_:
Whenever the axis in the MCU is ready for a new command,
bBusy is low.
This can be overwritten by setting bExecute to 0, which is
done in a "stop".

The ecmc is doing it differently, setting bExecute to 0 could
mean that the axis is still busy, because it is ramping down.

In order to distinguish these, ask for "sFeatures".

Add a new bit drvlocal.supported.bBusyOldStyle (which is not
an ideal name but merely reflecting the history).
In future, we may find a better name.