EthercatMC: New PV "-En" controls the amplifier
So far the amplifier could be turned on by using the .CNEN field
in the motor/axis record.
This is still possible as before, when the axis is created with the
AMPLIFIER_ON_FLAG_USING_CNEN bit in flags.

Now this is obsolete (but still working to be backwards compatible)

A new PV is used to control the amplifier:
Instead of using "caput IOC:m1.CNEN 1" "caput IOC:m1-En 1" should be used.

The major advantage is that the "wait" now works in pyEpics:
epics.caput(motor + '-En', 1, wait=True)

The PV is reflectiong the state in the hardware, so if the amplifier
changes state in the controller, the PV will change.

The test scripts will be updated in separate commits.