axisRecord.cc: Remove some local variables
Continue to refactor the doBackLashXXX() functions:
Some local variables can be removed, e.g. double bvel = pmr->bvel;
When needed, we can read it directly from the record.

Note: passing pmr->bvel to the devSupMoveXXX() can be removed,
the functions can retrieve it themself. But that will be another commit.