Remove NTM NTMF
The documentation of the NTM field said:
"If NTM is "YES", the motor is immediately stopped and given a motion
command to the new position".

Actually, there was no code for this, the movement did always complete.
The only code that was left was the stop at overshoot, which frequently
confused users of servo motors.

Remove the NTM/NTMF fields and the if "going into the wrong direction ?"
then stop code.

If needed new code can be added later.