EthercatMCAxis: Reset ErrorId on disconnect()
When the controller is powered off, it tends to set an ErrorID
indication problems with the EtherCAT bus. (at least with the
powersupplies that we use).
While this is the best message we have at this moment, the asyn
layer detects later that the TCP/IP communication is broken.
When this happens, reset the ErrorId to 0 and show
"CommunicationError" in the "-MsgTxt PV.