Parameters for "read only" soft limits from controller
When a motion controller defines soft limits (so called "read only",
we can read out the values and save them in the IOC.

Add asynParamFloat64 parameters in the parameter library
to store these values:
motorHighLimitRO_,pC_->motorLowLimitRO_

Add asynParamInt32 parameters "motorFlagsHighLimitRO_, &motorFlagsLowLimitRO_
which indicate if the values are valid.

Add members in the MotorStatus struct to forward these values to device support
and later to the axis record.

Allocate 2 bits in the MFLG field so that device support and EPICS clients
can read see that the controller has defined soft limits for this motor.

None of the new variables is used yet, this will be part of the next commits.