Post events when MFLG value changes
When the MFLG (Motor flags) was introduced, it was assumed that the
flags are set once at startup, and that's it.
Now when testing whith "switchable read-only" values like
MF_HIGH_LIMIT_RO (the controller has read-only soft high limt),
it turns out that a monitor event is missing when MFLG changes.

Add it.