axisRecord.cc: Use doHomeSetcdir()
doHome() did already set pmr->cdir depending if the commmand
send to the controller was HOME_FOR or HOME_REV.

However, the callers of doHome() set pmt->cdir on their own,
overwriting the value.

This is not ideal, to set a value from within one function, and letting
it being overwritten directly after the function returns.

Keep the setting of pmr->cdir in doHome(), and rename the function
into doHomeSetcdir().