axisRecord: Correct/improve setting of CDIR
When MRES < 0, the direction between dial- and raw- coordinates is inverted.
CDIR should reflect the status of the "raw commanded direction", but was
wrong when doing positioning with MRES < 0.
Once running into a limit switch, the motor could not be moved out using
the VAL field. These movements where blocked in the record.
Jogging out of the limit switch did work, though.

Create setCDIRfromRawMove() and setCDIRfromDialMove() and let them
calculare CDIR and set it.
While there, add a monitor (MARK_AUX).

On the long run, there are more problems:
- Once a limit switch is hit, the record should not even try to
  command a motion into the wrong direction (this is done for
  HOMF/HOMR, but nit for other movements)

- When a limit switch is hit, looking at CDIR may be wrong.
  Eithere it may be a third party move, or the motor may be
  overshooting.
  We should look at the traveling direction.
  And/or: What for the transition of the limit switch,
  and only send a stop when the limit switch gets active
  (it must have been passive before)