axisRecord.cc: set_user_high/lowlimit use set_dialXXXlimit()
Rewrite set_user_high/lowlimit() to use set_dial_highlimit() or
set_dial_lowlimit(), depending on DIR.

This simplies the code and allows to add the "range" check inside
set_dial_XXXlimit() in the next commit.
Do more code simplification when user limit is set, to prepare
for the "clipping" of the soft lints in the records to the values
of "read-only" software in the controller.