Improve asynAxisController::readGenericPointer()
This method used to retrive motor and encoder position from
the poller into a struct used in device support in axis record.

Both device support and the driver use already the same data structure,
"struct MotorStatus".
Today not all elements are copied, but e.g. the flags are missing.
Use memcpy to copy the whole struct.

This fixes a problem that the "read only" soft limits are known in
the record, but not the flags in pmr->MFLG which say the the values
are valid.
Whith this change the high- and low- softlimits are clipped (if needed) when
the record is initialized.