EthercatMC: Remove mres handling
In the beginning we started with a "dummy motor resolution", which
was set to 1 um.

The complete chain was to set EQU in the record to cm, set MRES to 0.001
so that the driver received um, and compensate for that in the driver
and convert um into mm when talking to the controller.

This was needed because device support truncated the motor/encoder
readback values to an integer value.
Since commit 7a125b9c, Make it possible to use encoder readback > 32 bit",
this is no longer needed: remove it.