TC: Switch off controller read only soft limits
When we try test that the limits switches can be reached, the soft
limits needs to be switched off.
Introduce a library function setSoftLimitsOff() and use it.

Since the introduction of "read only soft limits" in the controller,
these needs to be switched off. This works for the simulator and for
real EthercatMC systems (as long as the db file is loaded)