axisRecord; Remove doMoveRelDial(), doMoveAbsDial()
Refactor the code:
When we want to move to a dial-position, and external encoder (or EPICS PV)
is used for feedback, then all moves are relative.

For parts of the state machine, make this decision in
doMoveDialPosition(). (more refactoring is needed).

At the same time, ask doMoveDialPosition() to either use backlash-
or positioning paramters, so that we don't need to pass "velo" or "acc"
into this function, only the "moveMode", which is either moveModePosition
or moveModeBacklash.

This simplifies the code calling doMoveDialPosition().