EthercatMC: Accept MAIN or Main
The "main" program is typically called "MAIN", but the
driver expects "Main".
This has always been working with the EpicsComModule,
which is case-insensitive and send banck what the driver
asked for.

This does not work when the m-epics-twincat-ads bridge is used,
which replies with the name in the PLC.

Solution:
Compare "MAIN." case-insensitive before call sscanf().